TSK-1640: Configure taskana-spi-routing-dmn-router module properly
