+++ .github/workflows/continuous-integration.yml
@@ -236,6 +236,7 @@ jobs:
           - taskana-cdi-example
           - taskana-spring
           - taskana-spring-example
+          - taskana-spi-routing-dmn-router
           - taskana-rest-spring
           - taskana-rest-spring-example-common
           - taskana-loghistory-provider
@@ -369,7 +370,8 @@ jobs:
           :taskana-common-parent,:taskana-common,:taskana-common-security,:taskana-common-data,:taskana-common-test,\
           :taskana-lib-parent,:taskana-core,:taskana-cdi,:taskana-spring,\
           :taskana-rest-parent,:taskana-web,:taskana-rest-spring,\
-          :taskana-history-parent,:taskana-simplehistory-provider,:taskana-simplehistory-rest-spring,:taskana-loghistory-provider
+          :taskana-history-parent,:taskana-simplehistory-provider,:taskana-simplehistory-rest-spring,:taskana-loghistory-provider,\
+          :taskana-routing-parent,:taskana-spi-routing-dmn-router
         env:
           GPG_KEY_NAME: ${{ secrets.GPG_KEY_NAME }}
           GPG_PASSPHRASE: ${{ secrets.GPG_PASSPHRASE }}
