+++ common/taskana-common-security/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <artifactId>taskana-common-security</artifactId>
 
@@ -9,8 +9,8 @@
   <description>The global security principals needed for user and group authentication</description>
 
   <parent>
-    <artifactId>taskana-common-parent</artifactId>
     <groupId>pro.taskana</groupId>
+    <artifactId>taskana-common-parent</artifactId>
     <version>4.6.1-SNAPSHOT</version>
   </parent>
