+++ common/taskana-common/src/test/resources/pro/taskana/common/internal/util/fileInClasspath.txt
@@ -1,0 +1,1 @@
+This file is in the classpath
