+++ history/taskana-loghistory-provider/pom.xml
@@ -1,7 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <artifactId>taskana-loghistory-provider</artifactId>
+
+  <name>${project.groupId}:${project.artifactId}</name>
+  <description>The taskana history events logging plugin to include in your project.</description>
 
   <parent>
     <groupId>pro.taskana.history</groupId>
@@ -9,13 +14,6 @@
     <version>4.6.1-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <artifactId>taskana-loghistory-provider</artifactId>
-
-  <name>${project.groupId}:${project.artifactId}</name>
-  <description>The taskana history events logging plugin to include in your project.</description>
 
   <dependencies>
     <dependency>
