+++ lib/taskana-core/pom.xml
@@ -1,6 +1,6 @@
 <project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xmlns="http://maven.apache.org/POM/4.0.0"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+         xmlns="http://maven.apache.org/POM/4.0.0"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <artifactId>taskana-core</artifactId>
 
@@ -113,8 +113,9 @@
       <scope>test</scope>
     </dependency>
   </dependencies>
-  <!-- this repository is needed to fetch com.ibm.db2.jcc -->
+
   <repositories>
+    <!-- this repository is needed to fetch com.ibm.db2.jcc -->
     <repository>
       <id>novatec public</id>
       <name>novatec-repository</name>
