+++ lib/taskana-core/src/main/java/pro/taskana/classification/api/ClassificationQueryColumnName.java
@@ -31,7 +31,7 @@ public enum ClassificationQueryColumnName implements QueryColumnName {
   CUSTOM_7("custom_7"),
   CUSTOM_8("custom_8");
 
-  private String name;
+  private final String name;
 
   ClassificationQueryColumnName(String name) {
     this.name = name;
