--- qa/eclipse/taskana.importorder
@@ -1,5 +1,0 @@
-#Organize Import Order
-#Thu Dec 26 21:03:36 CET 2019
-2=pro.taskana
-1=
-0=\#
