+++ routing/taskana-spi-routing-dmn-router/src/test/resources/META-INF/services/pro.taskana.spi.routing.api.TaskRoutingProvider
@@ -1,0 +1,1 @@
+pro.taskana.routing.dmn.DmnTaskRouter
