+++ routing/taskana-spi-routing-dmn-router/src/test/resources/taskana.properties
@@ -1,0 +1,25 @@
+taskana.roles.user=cn=ksc-users,cn=groups,OU=Test,O=TASKANA | teamlead-1 | teamlead-2 | user-1-1 | user-1-2 | user-2-1 | user-2-2 | user-b-1 | user-b-2
+taskana.roles.admin=admin | uid=admin,cn=users,OU=Test,O=TASKANA
+taskana.roles.businessadmin=businessadmin | cn=business-admins,cn=groups,OU=Test,O=TASKANA
+taskana.roles.monitor=monitor | cn=monitor-users,cn=groups,OU=Test,O=TASKANA
+taskana.roles.taskadmin=taskadmin
+
+taskana.domains=DOMAIN_A,DOMAIN_B,DOMAIN_C
+taskana.classification.types=TASK,DOCUMENT
+taskana.classification.categories.task= EXTERNAL, manual, autoMAtic, Process
+taskana.classification.categories.document= EXTERNAL
+
+taskana.jobs.maxRetries=3
+taskana.jobs.batchSize=50
+taskana.jobs.cleanup.runEvery=P1D
+taskana.jobs.cleanup.firstRunAt=2018-07-25T08:00:00Z
+taskana.jobs.cleanup.minimumAge=P14D
+taskana.jobs.history.batchSize=50
+taskana.jobs.history.cleanup.firstRunAt=2018-07-25T08:00:00Z
+taskana.jobs.history.cleanup.minimumAge=P14D
+taskana.jobs.history.cleanup.runEvery=P1D
+taskana.german.holidays.enabled=true
+taskana.german.holidays.corpus-christi.enabled=true
+taskana.historylogger.name=AUDIT
+
+taskana.routing.dmn=/dmn-table.dmn
