--- taskana-routing-parent/pom.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <artifactId>taskana-routing-parent</artifactId>
-  <packaging>pom</packaging>
-
-  <name>${project.groupId}:${project.artifactId}</name>
-  <description>This pom is parent to all taskana routing modules</description>
-
-  <parent>
-    <groupId>pro.taskana</groupId>
-    <artifactId>taskana-parent</artifactId>
-    <version>4.6.1-SNAPSHOT</version>
-    <relativePath>../pom.xml</relativePath>
-  </parent>
-
-  <modules>
-    <module>taskana-spi-routing-dmn-router</module>
-  </modules>
-
-</project>
