--- taskana-routing-parent/taskana-spi-routing-dmn-router/src/test/resources/META-INF/services/pro.taskana.spi.routing.api.TaskRoutingProvider
@@ -1,1 +1,0 @@
-pro.taskana.routing.dmn.DmnTaskRouter
