--- taskana-routing-parent/taskana-spi-routing-dmn-router/src/test/resources/dmn-table.dmn
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<definitions xmlns:ns0="http://camunda.org/schema/1.0/dmn" xmlns="http://www.omg.org/spec/DMN/20151101/dmn.xsd" xmlns:biodi="http://bpmn.io/schema/dmn/biodi/1.0" id="definitions" name="definitions" namespace="http://camunda.org/schema/1.0/dmn" exporter="Camunda Modeler" exporterVersion="3.3.5">
-  <decision id="workbasketRouting" name="Workbasket Routing">
-    <extensionElements>
-      <biodi:bounds x="150" y="150" width="180" height="80" />
-    </extensionElements>
-    <decisionTable id="workbasketRouting_decisionTable" hitPolicy="FIRST">
-      <input id="workbasketRouting_decisionTable-input_2" label="porValue" ns0:inputVariable="input">
-        <inputExpression id="workbasketRouting_decisionTable-inputExpression_2" typeRef="string">
-          <text>task.primaryObjRef.value</text>
-        </inputExpression>
-      </input>
-      <input id="InputClause_1qkek3h" label="porType" ns0:inputVariable="input">
-        <inputExpression id="LiteralExpression_1k84ufl" typeRef="string">
-          <text>task.primaryObjRef.type</text>
-        </inputExpression>
-      </input>
-      <output id="workbasketRouting_decisionTable-output_workbasketKey" label="Workbasket key" name="workbasketKey" typeRef="string" />
-      <output id="workbasketRouting_decisionTable-output_domain" label="Domain" name="domain" typeRef="string" />
-      <rule id="workbasketRouting_decisionTable-rule_0">
-        <inputEntry id="workbasketRouting_decisionTable-rule_0-inputEntry_2">
-          <text>"00000001"</text>
-        </inputEntry>
-        <inputEntry id="UnaryTests_0n5ac31">
-          <text>"MyType1"</text>
-        </inputEntry>
-        <outputEntry id="workbasketRouting_decisionTable-rule_0-outputEntry_workbasketKey">
-          <text>"GPK_KSC"</text>
-        </outputEntry>
-        <outputEntry id="workbasketRouting_decisionTable-rule_0-outputEntry_domain">
-          <text>"DOMAIN_A"</text>
-        </outputEntry>
-      </rule>
-      <rule id="DecisionRule_0ua6ja6">
-        <inputEntry id="UnaryTests_1kdwhvv">
-          <text>"00000001"</text>
-        </inputEntry>
-        <inputEntry id="UnaryTests_03hbryc">
-          <text></text>
-        </inputEntry>
-        <outputEntry id="workbasketRouting_decisionTable-rule_1-outputEntry_workbasketKey">
-          <text>"GPK_KSC_1"</text>
-        </outputEntry>
-        <outputEntry id="workbasketRouting_decisionTable-rule_1-outputEntry_domain">
-          <text>"DOMAIN_A"</text>
-        </outputEntry>
-      </rule>
-      <rule id="DecisionRule_0elr8ov">
-        <inputEntry id="UnaryTests_0n7qv3k">
-          <text></text>
-        </inputEntry>
-        <inputEntry id="UnaryTests_0jwi2ra">
-          <text>"MyType1"</text>
-        </inputEntry>
-        <outputEntry id="workbasketRouting_decisionTable-rule_2-outputEntry_workbasketKey">
-          <text>"GPK_KSC_2"</text>
-        </outputEntry>
-        <outputEntry id="workbasketRouting_decisionTable-rule_2-outputEntry_domain">
-          <text>"DOMAIN_A"</text>
-        </outputEntry>
-      </rule>
-    </decisionTable>
-  </decision>
-</definitions>
